ingestion/requirements.txt
# ingestion/requirements.txt
spotipy==2.22.1
kafka-python==2.0.2